Begin to add type hinting to package.
Type hinting is mostly implemented for `pso_classes.py`, though it is not fully
up-to-date with the new `loss.py` layout. Type hinting is mostly added to the remaining
files, though there is a major issue that loss values can be a float or Tensor depending
on whether get_optim_mode() is on, and `mypy` doesn't realise that.